# src/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This api call data from youtube anf table",
        "title": "Dataset Collection API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "10.201.252.7:30014",
    "paths": {
...
        "/filterfun/trainTwOrg2TrainLpr/{youtubeId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lpr resualt"
                ],
                "summary": "trigger video into lpr detect image",
                "operationId": "train-twOrg-to-train-lpr",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Youtube ID",
                        "name": "youtubeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "We need Youtube ID!!",
                        "schema": {
                            "$ref": "#/definitions/string"
                        }
                    }
                }
            }
        },
        "/filterfun/trainTwOrg2TrainYolo/{youtubeId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "yolo resualt"
                ],
                "summary": "trigger video into yolo detect image",
                "operationId": "train-twOrg-to-train-yolo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Youtube Id",
                        "name": "youtubeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "We need Youtube ID!!",
                        "schema": {
                            "$ref": "#/definitions/string"
                        }
                    }
                }
            }
        },
...
        "/filterfun/youtubeUrl": {
            "post": {
                "description": "Download youtube video by url",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "get data list by filter parameter"
                ],
                "summary": "download youtube video by url",
                "operationId": "url-to-file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "filename",
                        "name": "filename",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "url",
                        "name": "url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "We need ID!!",
                        "schema": {
                            "$ref": "#/definitions/string"
                        }
                    },
                    "404": {
                        "description": "Can not find ID",
                        "schema": {
                            "$ref": "#/definitions/string"
                        }
                    }
                }
            }
        },
...
    }
}